+++ lib/Dialect/RTL/Dialect.cpp
@@ -6,8 +6,8 @@
 #include "circt/Dialect/RTL/Ops.h"
 #include "circt/Dialect/RTL/Types.h"
 #include "mlir/IR/Builders.h"
+#include "mlir/IR/BuiltinTypes.h"
 #include "mlir/IR/DialectImplementation.h"
-#include "mlir/IR/StandardTypes.h"
 
 using namespace circt;
 using namespace rtl;
@@ -87,20 +87,3 @@ Operation *RTLDialect::materializeConstant(OpBuilder &builder, Attribute value,
 
 // Provide implementations for the enums we use.
 #include "circt/Dialect/RTL/RTLEnums.cpp.inc"
-
-/// Parses a type registered to this dialect. Parse out the mnemonic then invoke
-/// the tblgen'd type parser dispatcher.
-Type RTLDialect::parseType(DialectAsmParser &parser) const {
-  llvm::StringRef mnemonic;
-  if (parser.parseKeyword(&mnemonic))
-    return Type();
-  return generatedTypeParser(getContext(), parser, mnemonic);
-}
-
-/// Print a type registered to this dialect. Try the tblgen'd type printer
-/// dispatcher then fail since all RTL types are defined via ODS.
-void RTLDialect::printType(Type type, DialectAsmPrinter &printer) const {
-  if (succeeded(generatedTypePrinter(type, printer)))
-    return;
-  llvm_unreachable("unexpected 'rtl' type");
-}
