+++ lib/Dialect/RTL/Types.cpp
@@ -5,12 +5,29 @@
 //===----------------------------------------------------------------------===//
 
 #include "circt/Dialect/RTL/Types.h"
+#include "mlir/IR/BuiltinTypes.h"
 #include "mlir/IR/DialectImplementation.h"
-#include "mlir/IR/StandardTypes.h"
 #include "llvm/ADT/TypeSwitch.h"
 
 using namespace mlir;
 using namespace circt::rtl;
+
+/// Parses a type registered to this dialect. Parse out the mnemonic then invoke
+/// the tblgen'd type parser dispatcher.
+Type RTLDialect::parseType(DialectAsmParser &parser) const {
+  llvm::StringRef mnemonic;
+  if (parser.parseKeyword(&mnemonic))
+    return Type();
+  return generatedTypeParser(getContext(), parser, mnemonic);
+}
+
+/// Print a type registered to this dialect. Try the tblgen'd type printer
+/// dispatcher then fail since all RTL types are defined via ODS.
+void RTLDialect::printType(Type type, DialectAsmPrinter &printer) const {
+  if (succeeded(generatedTypePrinter(type, printer)))
+    return;
+  llvm_unreachable("unexpected 'rtl' type");
+}
 
 /// Return true if the specified type can be used as an RTL value type, that is
 /// the set of types that can be composed together to represent synthesized,
